Merge origin/main and add validation to reject unsupported arguments
(sinks, sparse_mla_top_k, enable_pdl, skip_softmax_threshold_scale_factor)
in the cute-dsl backend with clear error messages.

+++ flashinfer/cute_dsl/rmsnorm_fp4quant.py
@@ -750,8 +750,8 @@ def _get_compiled_kernel(
             y_uint8,
             s_tensor,
             global_scale,
-            Int32(M),
-            Float32(eps),
+            M,
+            eps,
         )
 
     return tensor_api
